+++ functions/Start-PSTutorial.ps1
@@ -10,8 +10,9 @@ Function Start-PSTutorial {
         [string]$Topic
      )
 
-    $list = @"
-    $($PSStyle.Underline+$PSStyle.Foreground.Green)$($strings.menuTitle)$($PSStyle.Reset)
+     <#
+     $list = @"
+     #$($PSStyle.Underline+$PSStyle.Foreground.Green)$($strings.menuTitle)$($PSStyle.Reset)
 
     1 - $($strings.psEssentials)
     2 - Get-Command
@@ -20,21 +21,37 @@ Function Start-PSTutorial {
     5 - $($strings.quit)
 
 "@
+ #>
 
+#29 Nov 2025 use the private Format-BorderBox function to stylize the menu
+    [string[]]$list = @(
+"                    "
+"    1 - $($strings.psEssentials)    ",
+"    2 - Get-Command",
+"    3 - Get-Help",
+"    4 - Get-Member",
+"    5 - $($strings.quit)"
+"                        "
+)
     If ($PSBoundParameters.ContainsKey('Topic')) {
         #launch the tutorial from the module-scoped hashtable
         & $Tutorials[$Topic]
     }
     else {
         Clear-Host
-        $list
-        [int]$r = Read-Host "    $($strings.menuSelect) [1-5]"
+        Format-BorderBox -text $list -title $strings.menuTitle -BorderColor "`e[93m"
+        Try {
+            [int]$r = Read-Host "    $($PSStyle.Italic)$($PSStyle.Foreground.BrightGreen)$($strings.menuSelect)$($PSStyle.Reset) [1-5]"
 
-        Switch ($r) {
-            1 { &$tutorials['PowerShell Essentials'] -menu }`
-            2 { &$tutorials['Get-Command'] -menu}
-            3 { &$tutorials['Get-Help'] -menu }
-            4 { &$tutorials['Get-Member'] -menu}
-        }
+            Switch ($r) {
+                1 { &$tutorials['PowerShell Essentials'] -menu }`
+                2 { &$tutorials['Get-Command'] -menu}
+                3 { &$tutorials['Get-Help'] -menu }
+                4 { &$tutorials['Get-Member'] -menu}
+            }
+        } #Try
+        Catch {
+            #if the user enters any other value, consider it a quit or cancel
+        } #Catch
     }
 }